Complete refactoring - Spring IO Platform - HikariCP - Onion architecture - Spring Data
