+++ src/main/java/com/jeetemplates/app/interfaces/web/exceptions/DefaultExceptionHandlerFactory.java
@@ -1,0 +1,41 @@
+/**
+ * 
+ */
+package com.jeetemplates.app.interfaces.web.exceptions;
+
+import javax.faces.context.ExceptionHandler;
+import javax.faces.context.ExceptionHandlerFactory;
+
+/**
+ * Exception factory.
+ * 
+ * @author jeetemplates
+ */
+public class DefaultExceptionHandlerFactory extends ExceptionHandlerFactory {
+
+	/**
+	 * {@link ExceptionHandlerFactory}.
+	 */
+	private ExceptionHandlerFactory parent;
+
+	/**
+	 * Constructor
+	 * 
+	 * @param parent
+	 *            : parent
+	 */
+	public DefaultExceptionHandlerFactory(ExceptionHandlerFactory parent) {
+		this.parent = parent;
+	}
+
+	/**
+	 * @see javax.faces.context.ExceptionHandlerFactory#getExceptionHandler()
+	 */
+	@Override
+	public ExceptionHandler getExceptionHandler() {
+		ExceptionHandler eh = parent.getExceptionHandler();
+		eh = new DefaultExceptionHandler(eh);
+		return eh;
+	}
+
+}
