--- src/main/java/com/jeetemplates/app/persistence/HelloWorldDao.java
@@ -1,16 +1,0 @@
-/**
- * 
- */
-package com.jeetemplates.app.persistence;
-
-import com.jeetemplates.app.common.persistence.BusinessDao;
-import com.jeetemplates.app.domain.HelloWorld;
-
-/**
- * DAO of HelloWorld
- * 
- * @author jeetemplates
- */
-public interface HelloWorldDao extends BusinessDao<Long, HelloWorld> {
-
-}
