+++ src/main/resources/logback.xml
@@ -2,19 +2,19 @@
 
 <configuration>
 
-	<appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-		<encoder>
-			<pattern>%d %5p | %t | %-55logger{55} | %m %n</pattern>
-		</encoder>
-	</appender>
-	<logger name="com.jeetemplates">
-		<level value="INFO" />
-	</logger>
-	<logger name="org.springframework">
-		<level value="INFO" />
-	</logger>
-	<root>
-		<level value="INFO" />
-		<appender-ref ref="CONSOLE" />
-	</root>
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d %5p | %t | %-55logger{55} | %m %n</pattern>
+        </encoder>
+    </appender>
+    <logger name="com.jeetemplates">
+        <level value="INFO" />
+    </logger>
+    <logger name="org.springframework">
+        <level value="INFO" />
+    </logger>
+    <root>
+        <level value="INFO" />
+        <appender-ref ref="CONSOLE" />
+    </root>
 </configuration>
