--- src/main/resources/persistence/hibernate.properties
@@ -1,10 +1,0 @@
-# Hibernate properties
-hibernate.show_sql=false
-hibernate.format_sql=true
-hibernate.hbm2ddl.auto=update
-hibernate.dialect=org.hibernate.dialect.HSQLDialect
-hibernate.temp.use_jdbc_metadata_defaults=false
-hibernate.c3p0.min_size=5
-hibernate.c3p0.max_size=50
-hibernate.c3p0.max_statements=100
-hibernate.c3p0.idle_test_period=100
