--- src/test/java/com/jeetemplates/app/common/service/BaseServiceTest.java
@@ -1,17 +1,0 @@
-/**
- * 
- */
-package com.jeetemplates.app.common.service;
-
-import org.junit.runner.RunWith;
-import org.mockito.runners.MockitoJUnitRunner;
-
-/**
- * Base class for service test.
- * 
- * @author jeetemplates
- */
-@RunWith(MockitoJUnitRunner.class)
-public abstract class BaseServiceTest {
-
-}
