+++ src/test/resources/dataset/dataset.xml
@@ -2,7 +2,7 @@
 
 <dataset>
 	
-	<HELLO_WORLD ID="1" LAST_NAME="premier" FIRST_NAME="premier" />
-	<HELLO_WORLD ID="2" LAST_NAME="second" FIRST_NAME="second" />
+	<HelloWorld id="1" lastName="premier" firstName="premier" />
+	<HelloWorld id="2" lastName="second" firstName="second" />
 
 </dataset>
