--- src/test/resources/persistence/test-datasource.properties
@@ -1,5 +1,0 @@
-# Datasource properties
-jdbc.driverClass=org.hsqldb.jdbcDriver
-jdbc.url=jdbc:hsqldb:mem:testdb
-jdbc.user=sa
-jdbc.password=
