+++ src/test/resources/test-conf.properties
@@ -1,0 +1,14 @@
+# Datasource properties
+jdbc.driverClass=org.hsqldb.jdbcDriver
+jdbc.url=jdbc:hsqldb:mem:testdb
+jdbc.user=sa
+jdbc.password=
+datasource.pool.min_size=5
+datasource.pool.max_size=50
+
+# Hibernate properties
+hibernate.show_sql=true
+hibernate.format_sql=true
+hibernate.hbm2ddl.auto=update
+hibernate.dialect=org.hibernate.dialect.HSQLDialect
+hibernate.temp.use_jdbc_metadata_defaults=false
